requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
